UPSTREAM: 100523: refactor rest.FinishRequest function
